Completed hashing logic for routes, linked Login and Signup pages, added create table SQL script
