+++ database/auth/authDbInit.sql
@@ -1,0 +1,6 @@
+CREATE TABLE users (
+  id SERIAL PRIMARY KEY,
+  name VARCHAR(60) NOT NULL,
+  email VARCHAR(60) NOT NULL,
+  password VARCHAR(60) NOT NULL
+);
