+++ positive_charge/src/components/auth/Login.jsx
@@ -2,28 +2,21 @@ import React from 'react';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
 
-class Signup extends React.Component {
+
+class Login extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      name: '',
       email: '',
       password: ''
     }
-    this.handleChange = this.handleChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(e) {
-    this.setState({
-      [e.target.name]: e.target.value
-    })
+    this.handleChange = this.handleChange.bind(this);
   }
 
   handleSubmit(e) {
     e.preventDefault();
-    axios.post('/signup', {
-      name: this.state.name,
+    axios.post('/login', {
       email: this.state.email,
       password: this.state.password
     })
@@ -35,15 +28,17 @@ class Signup extends React.Component {
     })
   }
 
-  render () {
+  handleChange(e) {
+    this.setState({
+      [e.target.name]: e.target.value
+    })
+  }
+
+  render() {
     return (
-      <div className="signup">
-        <h1>Sign Up</h1>
+      <div className="login">
+        <h1>Login</h1>
         <form onSubmit={this.handleSubmit}>
-          <div>
-            <label htmlFor="name">Full Name: </label>
-            <input type="text" name="name" onChange={this.handleChange} required />
-          </div>
           <div>
             <label htmlFor="email">Email: </label>
             <input type="email" name="email" onChange={this.handleChange} required />
@@ -52,14 +47,14 @@ class Signup extends React.Component {
             <label htmlFor="password">Password: </label>
             <input type="text" name="password" onChange={this.handleChange} required />
           </div>
-          <input type="submit" value="Sign Up" />
+          <input type="submit" value="Login" />
         </form>
-        <div>Already have an account?
-          <Link to="/login">Login</Link>
+        <div>Don't have an account?
+          <Link to="/signup">Sign Up</Link>
         </div>
       </div>
     )
   }
 }
 
-export default Signup;
+export default Login;
